moq-clock: convert to a moq-native example
Replaces the standalone moq-clock crate with rs/moq-native/examples/clock.rs,
joining the existing chat example. The clock binary was already marked
publish = false on crates.io and had no external consumers; folding it
into the examples folder keeps it building with the workspace, drops a
member from `cargo check`'s target set, and removes the release-plz
config and Nix derivation that existed only to build a binary nobody
shipped.

Also fixes the release-plz CI failure on main: with the crate gone, the
`publish = false` (Cargo.toml) vs `publish = true` (release-plz default)
mismatch that aborted `release-plz release` no longer exists, so the
`.release-plz.toml` package block is dropped entirely.

Run with `cargo run -p moq-native --example clock -- --url ... --broadcast clock {publish,subscribe}`.
