--- rs/moq-clock/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "moq-clock"
-description = "CLOCK over QUIC"
-authors = ["Luke Curley"]
-repository = "https://github.com/moq-dev/moq"
-license = "MIT OR Apache-2.0"
-
-version = "0.10.21"
-edition = "2024"
-rust-version.workspace = true
-
-# Example binary, not a distributed crate. Stays in the workspace so it
-# builds with everything else, but never pushed to crates.io.
-publish = false
-
-keywords = ["quic", "http3", "webtransport", "media", "live"]
-categories = ["multimedia", "network-programming", "web-programming"]
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = { version = "1", features = ["backtrace"] }
-chrono = "0.4"
-clap = { version = "4", features = ["derive"] }
-moq-native = { workspace = true, default-features = true }
-moq-net = { workspace = true }
-tokio = { workspace = true, features = ["full"] }
-tracing = "0.1"
-url = "2"
